Move Object\Profile to Model\Profile
