+++ src/Protocol/Diaspora.php
@@ -17,9 +17,9 @@ use Friendica\Core\PConfig;
 use Friendica\Core\Worker;
 use Friendica\Database\DBM;
 use Friendica\Model\GlobalContact;
+use Friendica\Model\Profile;
 use Friendica\Network\Probe;
 use Friendica\Object\Contact;
-use Friendica\Object\Profile;
 use Friendica\Util\XML;
 
 use dba;
