+++ src/Protocol/PortableContact.php
@@ -13,8 +13,8 @@ use Friendica\Core\Config;
 use Friendica\Core\Worker;
 use Friendica\Database\DBM;
 use Friendica\Model\GlobalContact;
+use Friendica\Model\Profile;
 use Friendica\Network\Probe;
-use Friendica\Object\Profile;
 use dba;
 use DOMDocument;
 use DomXPath;
